requirements.txt
gspread==6.1.4
oauth2client==4.1.3
pandas==2.2.3
typer==0.15.1
python-telegram-bot==21.10
openai-whisper==20240930
ffmpeg-python==0.2.0